LayoutTests/sputnik/Conformance/12_Statement/12.1_Block/S12.1_A4_T1-expected.txt

S12.1_A4_T1

PASS TypeError: {__func} is not a function. (In '{__func}()', '{__func}' is an instance of Object)

TEST COMPLETE
